Ex 4.21 the user needed to delete a blog

+++ part4_blog_list_app/controllers/blogs.js
@@ -39,9 +39,30 @@ blogsRouter.post('/', async (request, response) => {
     response.status(200).json(savedBlog)
 })
 
+// blogsRouter.delete('/:id', async (request, response) => {
+//   await Blog.findByIdAndRemove(request.params.id)
+//   response.status(204).end()
+// })
+
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {    
+    return response.status(401).json({ 
+      error: 'token missing or invalid' 
+    })  
+  }
+
+  //if ( blog.user.id.toString() === decodedToken.id ){
+    if ( blog.user.toString() === decodedToken.id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+  
+  response.status(400).json({ 
+    error: 'not authorized to delete' 
+  })  
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
